requirement.txt
langchain==0.3.0
langchain-ibm
langchain-community
langchain-text-splitters
langchain-huggingface
faiss-cpu
pypdf
sentence-transformers
python-dotenv
streamlit